restructured project packages
