+++ app/src/main/java/com/example/greehousecontroller/data/api/HumidityApi.java
@@ -3,6 +3,11 @@ package com.example.greehousecontroller.data.api;
 
 
 
+import com.example.greehousecontroller.data.model.Humidity;
+
+import java.util.List;
+
+import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
