+++ app/src/main/java/com/example/greehousecontroller/data/repository/UserRepository.java
@@ -4,6 +4,7 @@ import android.app.Application;
 
 import androidx.annotation.NonNull;
 
+import com.example.greehousecontroller.data.model.UserLiveData;
 import com.firebase.ui.auth.AuthUI;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
