+++ app/src/main/java/com/example/greehousecontroller/ui/viewmodel/EditPotViewModel.java
@@ -4,8 +4,9 @@ import android.app.Application;
 
 import androidx.lifecycle.AndroidViewModel;
 
-import com.example.greehousecontroller.data.repository.PotRepository.PotRepository;
+
 import com.example.greehousecontroller.data.model.Pot;
+import com.example.greehousecontroller.data.repository.PotRepository;
 
 public class EditPotViewModel extends AndroidViewModel {
